requirements.txt
Django==2.0.1
django-bootstrap4==0.0.5
django-braces==1.12.0
pytz==2017.3
